fix(wrap): consume the explicit client argument separator
