--- .github/workflows/fix-wrap-separator-and-validate.yml
@@ -1,113 +1,0 @@
-name: Fix Wrap Separator and Validate
-
-on:
-  pull_request:
-    branches: [main]
-
-permissions:
-  contents: write
-
-jobs:
-  validate:
-    if: github.head_ref == 'feat/agent-compatibility-matrix' && github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    timeout-minutes: 20
-    steps:
-      - uses: actions/checkout@v6
-        with:
-          ref: ${{ github.head_ref }}
-          fetch-depth: 0
-          token: ${{ secrets.GITHUB_TOKEN }}
-      - uses: actions/setup-python@v6
-        with:
-          python-version: '3.12'
-      - uses: actions/setup-node@v6
-        with:
-          node-version: '22'
-      - name: Apply fail-closed separator fix
-        run: python .github/fix_wrap_separator.py
-      - name: Install product and pinned Copilot CLI
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install -e . pytest
-          npm install --global '@github/copilot@1.0.74'
-      - name: Run focused wrapper contracts
-        run: |
-          python -m compileall -q entroly/cli.py entroly/session_attach.py
-          python -m pytest -q tests/test_agent_compatibility.py tests/test_cli.py tests/test_cli_audit.py tests/test_session_attach.py
-      - name: Canonicalize exploratory Copilot prompt ordering
-        run: |
-          python - <<'PY'
-          from pathlib import Path
-
-          path = Path('.github/copilot_composite_e2e_probe.py')
-          lines = path.read_text(encoding='utf-8').splitlines()
-          start = next(
-              index
-              for index in range(len(lines) - 1)
-              if lines[index].strip() == '"-s",' and lines[index + 1].strip() == '"-p",'
-          )
-          end = next(
-              index
-              for index in range(start, len(lines))
-              if lines[index].strip() == '"--no-remote-export",'
-          )
-          indent = lines[start][: len(lines[start]) - len(lines[start].lstrip())]
-          lines[start : end + 1] = [
-              f'{indent}"-s",',
-              f'{indent}f"--allow-tool={{server_name}}",',
-              f'{indent}"--no-color",',
-              f'{indent}"--no-ask-user",',
-              f'{indent}"--no-auto-update",',
-              f'{indent}"--no-custom-instructions",',
-              f'{indent}"--disable-builtin-mcps",',
-              f'{indent}"--no-remote",',
-              f'{indent}"--no-remote-export",',
-              f'{indent}f"--prompt=Use the get_stats tool from MCP server {{server_name}} exactly once. After the tool succeeds, return exactly E2E_OK.",',
-          ]
-          path.write_text('\n'.join(lines) + '\n', encoding='utf-8')
-          PY
-      - name: Run real Copilot proxy and MCP round trip
-        env:
-          ENTROLY_DISABLE_UPDATE_CHECK: '1'
-        run: |
-          python .github/copilot_composite_e2e_probe.py
-          python - <<'PY'
-          import json
-          from pathlib import Path
-
-          report = json.loads(Path('copilot-composite-e2e.json').read_text())
-          if not report.get('passed'):
-              raise SystemExit(
-                  'Copilot composite E2E failed: ' +
-                  json.dumps({
-                      'classification': report.get('classification'),
-                      'error': report.get('error'),
-                      'requests_total': report.get('stats', {}).get('requests_total'),
-                      'grant_use_count': report.get('grant_use_count'),
-                      'advertised_get_stats_tools': report.get('advertised_get_stats_tools'),
-                      'tool_result_requests': report.get('tool_result_requests'),
-                      'invocation': report.get('invocation'),
-                  }, indent=2)
-              )
-          PY
-      - name: Upload E2E evidence
-        if: always()
-        uses: actions/upload-artifact@v6
-        with:
-          name: wrap-separator-copilot-e2e
-          path: |
-            copilot-composite-e2e.json
-            copilot-composite-e2e.log
-          if-no-files-found: warn
-          retention-days: 14
-      - name: Commit product-only fix
-        run: |
-          rm -f .github/fix_wrap_separator.py
-          rm -f .github/workflows/fix-wrap-separator-and-validate.yml
-          git checkout -- .github/copilot_composite_e2e_probe.py
-          git config user.name juyterman1000
-          git config user.email 208309368+juyterman1000@users.noreply.github.com
-          git add entroly/cli.py tests/test_agent_compatibility.py .github/fix_wrap_separator.py .github/workflows/fix-wrap-separator-and-validate.yml
-          git commit -m "fix(wrap): consume the explicit client argument separator"
-          git push origin HEAD:${{ github.head_ref }}
